+++ R/coxDKsplsDR.default.R
@@ -2,6 +2,7 @@
 #' @export
 
 coxDKsplsDR.default <- function(Xplan,time,time2,event,type,origin,typeres="deviance", collapse, weighted, scaleX=TRUE, scaleY=TRUE, ncomp=min(7,ncol(Xplan)), modepls="regression", plot=FALSE, allres=FALSE, eta, trace=FALSE,kernel="rbfdot",hyperkernel, verbose=TRUE,...) {
+dataX <- Xplan
 if(scaleX){Xplan <- scale(Xplan); XplanScal <- attr(Xplan,"scaled:scale"); XplanCent <- attr(Xplan,"scaled:center"); Xplan <- as.data.frame(Xplan)} else {Xplan <- as.data.frame(Xplan);XplanScal <- rep(1,ncol(Xplan)); XplanCent <- rep(0,ncol(Xplan))}
 if((scaleY & missing(time2))){time <- scale(time)}
 try(attachNamespace("survival"),silent=TRUE)
@@ -128,8 +129,10 @@ cox_DKsplsDR$call$data <- as.name("tt_DKsplsDR")
 CoeffCFull[,iii] <- c(cox_DKsplsDR$coefficients,rep(NA,ncomp-iii))
 }
 }
-return(list(tt_DKsplsDR=tt_DKsplsDR, cox_DKsplsDR=cox_DKsplsDR, DKsplsDR_mod=DKsplsDR_mod, DKsplsDR_modplsr=DKsplsDR_modplsr, XplanScal=XplanScal, XplanCent=XplanCent, CoeffCFull=CoeffCFull, kernDKsplsDR_mod=kernDKsplsDR_mod))}
+res <- list(tt_DKsplsDR=tt_DKsplsDR, cox_DKsplsDR=cox_DKsplsDR, DKsplsDR_mod=DKsplsDR_mod, DKsplsDR_modplsr=DKsplsDR_modplsr, XplanScal=XplanScal, XplanCent=XplanCent, CoeffCFull=CoeffCFull, kernDKsplsDR_mod=kernDKsplsDR_mod, dataX=dataX, RepY=YCsurv)
+if(exists("XplanFormula")){res$XplanFormula <- XplanFormula}
+class(res) <- c("coxDKsplsDRmodel", "list")
+return(res)}
 }
-