+++ R/coxplsDR.default.R
@@ -2,6 +2,7 @@
 #' @export
 
 coxplsDR.default <- function(Xplan,time,time2,event,type,origin,typeres="deviance", collapse, weighted, scaleX=TRUE, scaleY=TRUE, ncomp=min(7,ncol(Xplan)), modepls="regression", plot=FALSE, allres=FALSE,...) {
+dataX <- Xplan
 if(scaleX){Xplan <- scale(Xplan); XplanScal <- attr(Xplan,"scaled:scale"); XplanCent <- attr(Xplan,"scaled:center"); Xplan <- as.data.frame(Xplan)} else {Xplan <- as.data.frame(Xplan);XplanScal <- rep(1,ncol(Xplan)); XplanCent <- rep(0,ncol(Xplan))}
 if((scaleY & missing(time2))){time <- scale(time)}
 try(attachNamespace("survival"),silent=TRUE)
@@ -88,6 +89,9 @@ cox_plsDR$call$data <- as.name("tt_plsDR")
 CoeffCFull[,iii] <- c(cox_plsDR$coefficients,rep(NA,ncomp-iii))
 }
 }
-return(list(tt_plsDR=tt_plsDR, cox_plsDR=cox_plsDR, plsDR_mod=plsDR_mod, XplanScal=XplanScal, XplanCent=XplanCent, CoeffCFull=CoeffCFull))}
+res <- list(tt_plsDR=tt_plsDR, cox_plsDR=cox_plsDR, plsDR_mod=plsDR_mod, XplanScal=XplanScal, XplanCent=XplanCent, CoeffCFull=CoeffCFull, dataX=dataX, RepY=YCsurv)
+if(exists("XplanFormula")){res$XplanFormula <- XplanFormula}
+class(res) <- c("coxplsDRmodel", "list")
+return(res)}
 
 }