+++ NEWS.md
@@ -1,6 +1,6 @@
 # plsRcox 1.8.2
-* Fix issue with predict expected data handling
-
+* Fixed issue with predict expected data handling
+* Extended prediction support DR/sPLS paths and fixed related issues
 
 # plsRcox 1.8.1
 * Fix package as requested by CRAN